backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Redis==0.4.0
Flask-Limiter==3.5.0
marshmallow==3.20.1
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.1
plotly==5.17.0
bcrypt==4.0.1
PyJWT==2.8.0
redis==5.0.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
gunicorn==21.2.0
xlsxwriter==3.1.2 